requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-fireworks>=0.2.0
langchain-core>=0.3.0

# HTTP and async
httpx>=0.27.0
aiohttp>=3.9.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment and config
python-dotenv>=1.0.0

# Utilities
tiktoken>=0.7.0
python-dateutil>=2.8.0

# CLI
rich>=13.0.0
inquirer>=3.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
langchain-google-genai==2.1.12